Refactor entity folder
