+++ src/com/covid/vaccination/center/booking/command/AddCapacityCommand.java
@@ -1,11 +1,11 @@
 package com.covid.vaccination.center.booking.command;
 
-import com.covid.vaccination.center.booking.entity.CenterCapacity;
+import com.covid.vaccination.center.booking.dto.request.CenterCapacity;
 
 public class AddCapacityCommand implements ICommand {
     @Override
     public <T> T create(String[] input) {
-        return (T) new CenterCapacity(input[1],input[2],input[3]);
+        return (T) new CenterCapacity(input[1], input[2], input[3]);
     }
 
     @Override
