+++ src/com/covid/vaccination/center/booking/command/AddUserCommand.java
@@ -1,6 +1,6 @@
 package com.covid.vaccination.center.booking.command;
 
-import com.covid.vaccination.center.booking.entity.User;
+import com.covid.vaccination.center.booking.dto.request.User;
 import com.covid.vaccination.center.booking.service.UserService;
 
 public class AddUserCommand implements ICommand {
@@ -10,7 +10,7 @@ public class AddUserCommand implements ICommand {
 
     @Override
     public <T> T create(String[] input) {
-        return (T) new User(input[1],input[2],input[3],Integer.parseInt(input[4]),input[5],input[6]);
+        return (T) new User(input[1], input[2], input[3], Integer.parseInt(input[4]), input[5], input[6]);
     }
 
     @Override
