+++ src/com/covid/vaccination/center/booking/command/AddVaccinationCenterCommand.java
@@ -1,6 +1,6 @@
 package com.covid.vaccination.center.booking.command;
 
-import com.covid.vaccination.center.booking.entity.VaccinationCenter;
+import com.covid.vaccination.center.booking.dto.request.VaccinationCenter;
 import com.covid.vaccination.center.booking.service.AdminService;
 
 public class AddVaccinationCenterCommand implements ICommand {
