+++ src/com/covid/vaccination/center/booking/command/CancelBookingCommand.java
@@ -1,11 +1,11 @@
 package com.covid.vaccination.center.booking.command;
 
-import com.covid.vaccination.center.booking.entity.CancelVaccination;
+import com.covid.vaccination.center.booking.dto.request.CancelVaccination;
 
 public class CancelBookingCommand implements ICommand {
     @Override
     public <T> T create(String[] input) {
-        return (T) new CancelVaccination(input[1],input[2],input[3]);
+        return (T) new CancelVaccination(input[1], input[2], input[3]);
     }
 
     @Override
