+++ src/com/covid/vaccination/center/booking/service/UserService.java
@@ -1,6 +1,6 @@
 package com.covid.vaccination.center.booking.service;
 
-import com.covid.vaccination.center.booking.entity.User;
+import com.covid.vaccination.center.booking.dto.request.User;
 import com.covid.vaccination.center.booking.exceptions.UserAlreadyRegistered;
 
 import java.util.HashMap;
@@ -8,12 +8,12 @@ import java.util.Map;
 
 public class UserService {
 
-    Map<String,User> users = new HashMap<>();
+    Map<String, User> users = new HashMap<>();
 
-    public void addUser(User user){
+    public void addUser(User user) {
         String userId = user.getUserId();
-        if(users.containsKey(userId))
+        if (users.containsKey(userId))
             throw new UserAlreadyRegistered("User Already Exists");
-        users.put(userId,user);
+        users.put(userId, user);
     }
 }
